+++ nodes/Autotask/operations/base/getFieldInfo.operation.ts
@@ -3,16 +3,15 @@ import { BaseOperation } from './base-operation';
 import { OperationType } from '../../types/base/entity-types';
 import { handleErrors } from '../../helpers/errorHandler';
 import { getEntityMetadata } from '../../constants/entities';
-import type { IAutotaskEntity } from '../../types/base/entity-types';
-import type { IEntityField, IEntityInfo } from '../../types/base/entities';
+import type { IEntityField } from '../../types/base/entities';
 import { ERROR_TEMPLATES } from '../../constants/error.constants';
-import { buildEntityUrl } from '../../helpers/http/request';
-import { autotaskApiRequest } from '../../helpers/http';
+import { getEntityInfo, getFields } from '../../helpers/entity/api';
+import type { IUdfFieldDefinition } from '../../types/base/udf-types';
 
 /**
  * Operation to retrieve field information including standard and UDF fields
  */
-export class GetFieldInfoOperation<T extends IAutotaskEntity> extends BaseOperation {
+export class GetFieldInfoOperation extends BaseOperation {
 	constructor(
 		entityType: string,
 		context: IExecuteFunctions,
@@ -25,7 +24,7 @@ export class GetFieldInfoOperation<T extends IAutotaskEntity> extends BaseOperat
 	/**
 	 * Execute the operation
 	 */
-	public async execute(itemIndex: number): Promise<INodeExecutionData> {
+	public async execute(): Promise<INodeExecutionData> {
 		return handleErrors(
 			this.context,
 			async () => {
@@ -40,45 +39,19 @@ export class GetFieldInfoOperation<T extends IAutotaskEntity> extends BaseOperat
 					);
 				}
 
-				// Get entity info first to check for UDF support
-				const baseEndpoint = buildEntityUrl(this.entityType).replace(/\/$/, '');
-				const entityInfoResponse = await autotaskApiRequest.call(
-					this.context,
-					'GET',
-					`${baseEndpoint}/entityInformation`,
-				) as { info: IEntityInfo };
+				// Get entity info using cached function
+				const entityInfo = await getEntityInfo(this.entityType, this.context);
 
-				if (!entityInfoResponse?.info) {
-					throw new Error(
-						ERROR_TEMPLATES.validation
-							.replace('{type}', 'ValidationError')
-							.replace('{entity}', this.entityType)
-							.replace('{details}', 'Invalid entity info response format'),
-					);
-				}
+				// Get standard fields using cached function
+				const standardFieldsRaw = await getFields(this.entityType, this.context, { fieldType: 'standard' }) as IEntityField[];
+				const standardFields = this.processStandardFields(standardFieldsRaw);
 
-				const entityInfo = entityInfoResponse.info;
-
-				// Get standard fields from API
-				const standardFieldsResponse = await autotaskApiRequest.call(
-					this.context,
-					'GET',
-					`${baseEndpoint}/entityInformation/fields`,
-				) as { fields: IEntityField[] };
-
-				// Process standard fields
-				const standardFields = this.processStandardFields(standardFieldsResponse.fields || []);
-
-				// Get UDF fields if supported
+				// Get UDF fields if supported using cached function
 				let udfFields: IEntityField[] = [];
 				if (entityInfo.hasUserDefinedFields) {
-					console.debug(`[${new Date().toISOString()}] Entity ${this.entityType} supports UDF fields, fetching from: ${baseEndpoint}/entityInformation/userDefinedFields`);
-					const udfResponse = await autotaskApiRequest.call(
-						this.context,
-						'GET',
-						`${baseEndpoint}/entityInformation/userDefinedFields`,
-					) as { fields: IEntityField[] };
-					udfFields = this.processUdfFields(udfResponse.fields || []);
+					console.debug(`[${new Date().toISOString()}] Entity ${this.entityType} supports UDF fields, fetching UDF fields`);
+					const udfFieldsRaw = await getFields(this.entityType, this.context, { fieldType: 'udf' }) as IUdfFieldDefinition[];
+					udfFields = this.processUdfFields(udfFieldsRaw as unknown as IEntityField[]);
 					console.debug(`[${new Date().toISOString()}] Retrieved ${udfFields.length} UDF fields for ${this.entityType}`);
 				} else {
 					console.debug(`[${new Date().toISOString()}] Entity ${this.entityType} does not support UDF fields (hasUserDefinedFields: ${entityInfo.hasUserDefinedFields})`);