+++ nodes/Autotask/resources/ticketHistories/index.ts
@@ -1,0 +1,2 @@
+export { executeTicketHistoryOperation } from './execute';
+export { ticketHistoryFields } from './description';