+++ nodes/Autotask/resources/ticketNotes/description.ts
@@ -1,6 +1,6 @@
 import type { INodeProperties } from 'n8n-workflow';
 
-export const resourceFields: INodeProperties[] = [
+export const ticketNoteFields: INodeProperties[] = [
 	{
 		displayName: 'Operation',
 		name: 'operation',
@@ -9,51 +9,57 @@ export const resourceFields: INodeProperties[] = [
 		displayOptions: {
 			show: {
 				resource: [
-					'resource',
+					'ticketNote',
 				],
 			},
 		},
 		options: [
 			{
-				name: 'Get',
-				value: 'get',
-				description: 'Get a resource by ID',
-				action: 'Get a resource',
-			},
-			{
-				name: 'Get Many',
-				value: 'getMany',
-				description: 'Get multiple resources',
-				action: 'Get multiple resources',
+				name: 'Create',
+				value: 'create',
+				description: 'Create a ticket note',
+				action: 'Create a ticket note',
 			},
 			{
 				name: 'Update',
 				value: 'update',
-				description: 'Update a resource',
-				action: 'Update a resource',
+				description: 'Update a ticket note',
+				action: 'Update a ticket note',
+			},
+			{
+				name: 'Get',
+				value: 'get',
+				description: 'Get a ticket note by ID',
+				action: 'Get a ticket note',
+			},
+			{
+				name: 'Get Many',
+				value: 'getMany',
+				description: 'Get multiple ticket notes',
+				action: 'Get multiple ticket notes',
 			},
 			{
 				name: 'Count',
 				value: 'count',
-				description: 'Count number of resources',
-				action: 'Count resources',
+				description: 'Count number of ticket notes',
+				action: 'Count ticket notes',
 			},
 		],
-		default: 'get',
+		default: 'create',
 	},
 	{
-		displayName: 'Resource ID',
+		displayName: 'Note ID',
 		name: 'id',
 		type: 'string',
 		required: true,
 		default: '',
 		displayOptions: {
 			show: {
-				resource: ['resource'],
+				resource: ['ticketNote'],
 				operation: ['update', 'get'],
 			},
 		},
-
+		description: 'The ID of the note to update',
 	},
 	{
 		displayName: 'Fields',
@@ -66,8 +72,8 @@ export const resourceFields: INodeProperties[] = [
 		required: true,
 		displayOptions: {
 			show: {
-				resource: ['resource'],
-				operation: ['update', 'getMany', 'count'],
+				resource: ['ticketNote'],
+				operation: ['create', 'update', 'getMany', 'count'],
 			},
 		},
 		typeOptions: {